Change block count & block index

+++ src/block_iter.rs
@@ -32,18 +32,18 @@ where
         cluster: Cluster,
         block_index: Option<BlockIndex>,
     ) -> Self {
-        let blocks_per_cluster = fs.boot_record.blocks_per_cluster() as usize;
+        let blocks_per_cluster = u64::from(fs.boot_record.blocks_per_cluster());
 
         let (cluster, block_index) = if let Some(block_index) = block_index {
-            let cluster_offset = block_index.0 / blocks_per_cluster as u32;
-            let block_index = BlockIndex(block_index.0 % blocks_per_cluster as u32);
-            (Cluster(cluster.0 + cluster_offset), Some(block_index))
+            let cluster_offset = block_index.0 / blocks_per_cluster;
+            let block_index = BlockIndex(block_index.0 % blocks_per_cluster);
+            (Cluster(cluster.0 + cluster_offset as u32), Some(block_index))
         } else {
             (cluster, block_index)
         };
 
         BlockIndexClusterIter {
-            counter: blocks_per_cluster,
+            counter: blocks_per_cluster as usize,
             cluster_iter: FatClusterIter::new(fs, cluster),
             block_index,
             last_cluster: None,
